requirements.txt
Flask==2.3.2
pandas==1.5.3
selenium==4.11.3
selenium-wire==4.10.1
webdriver-manager==3.8.6
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
python-anticaptcha==1.0.0
fake-useragent==1.2.0
openai==0.27.4
concurrent-log-handler==0.9.19
